+++ .travis.yml
@@ -1,0 +1,10 @@
+language: python
+
+python:
+  - "3.6"
+
+script:
+  - coverage run -m unittest discover -s test -p 'test_*.py'
+
+after_success:
+- codecov
