+++ requirements.txt
@@ -1,2 +1,1 @@
-numpy
 pyglet
